+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/language/LanguageActivity.kt
@@ -54,9 +54,9 @@ class LanguageActivity : CommonTitleActivity() {
         when (view.id) {
             R.id.languageApp -> ToastUtils.showLong(R.string.language)
             R.id.languageActivity -> ToastUtils.showLong(getString(R.string.language))
-            R.id.applySimpleChineseLanguage -> LanguageUtils.applyLanguage(Locale.SIMPLIFIED_CHINESE, null)
-            R.id.applyAmericanLanguage -> LanguageUtils.applyLanguage(Locale.US, null)
-            R.id.applySystemLanguage -> LanguageUtils.applySystemLanguage(null)
+            R.id.applySimpleChineseLanguage -> LanguageUtils.applyLanguage(Locale.SIMPLIFIED_CHINESE, "com.blankj.launcher.pkg.MainActivity")
+            R.id.applyAmericanLanguage -> LanguageUtils.applyLanguage(Locale.US, "")
+            R.id.applySystemLanguage -> LanguageUtils.applySystemLanguage("")
         }
     }
 }